+++ app/src/main/java/com/orienteering/handrail/services/CourseService.kt
@@ -20,5 +20,5 @@ interface CourseService {
     fun readAll(): Call<List<Course>>
 
     @GET("users/{id}/courses")
-    fun readAllByUser(@Path("id")userID : Int): Call<List<Course>>
+    fun readAllByUser(@Path("id")userID : Long): Call<List<Course>>
 }
